recruitcrm/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
boto3==1.35.76
botocore==1.35.76
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
Django==5.1.4
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
numpy==2.2.1
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
PyQt3D==5.15.6
PyQt5==5.15.10
PyQt5_sip==12.13.0
PyQt6==6.7.1
PyQt6-3D==6.7.0
PyQt6-Charts==6.7.0
PyQt6-DataVisualization==6.7.0
PyQt6-NetworkAuth==6.7.0
PyQt6-WebEngine==6.7.0
PyQt6_sip==13.8.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
s3transfer==0.10.4
six==1.17.0
SQLAlchemy==2.0.35
sqlparse==0.5.3
TBB==0.2
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.0.4
wheel==0.45.1
zipp==3.20.2
python-decouple
pyrebase4
firebase-admin
django-cors-headers
whitenoise
drf-yasg
